% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTBLS1.R
\name{MTBLS1}
\alias{MTBLS1}
\title{Spectra dataset and ppm of MTBLS1 dataset.}
\usage{
MTBLS1()
}
\value{
Spectra dataset and ppm of MTBLS1 dataset.
}
\description{
Spectra dataset and ppm of MTBLS1 dataset.
}
\examples{
data=MTBLS1()
}
